+++ src/main/java/com/artillexstudios/axvaults/commands/subcommands/ReplaceItems.java
@@ -3,7 +3,6 @@ package com.artillexstudios.axvaults.commands.subcommands;
 import com.artillexstudios.axvaults.converters.VaultItemReplacer;
 import org.bukkit.command.CommandSender;
 
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
 import static com.artillexstudios.axvaults.AxVaults.MESSAGEUTILS;
@@ -13,9 +12,7 @@ public enum ReplaceItems {
 
     public void execute(CommandSender sender) {
         CompletableFuture.runAsync(() -> {
-            int updatedVaults = new VaultItemReplacer().replaceItemsInVaults();
-            MESSAGEUTILS.sendLang(sender, "replace-items.finished",
-                    Map.of("%updated%", String.valueOf(updatedVaults)));
+            new VaultItemReplacer().run();
         });
         MESSAGEUTILS.sendLang(sender, "replace-items.started");
     }
