+++ src/main/java/com/artillexstudios/axvaults/database/impl/MySQL.java
@@ -2,7 +2,6 @@ package com.artillexstudios.axvaults.database.impl;
 
 import com.artillexstudios.axapi.serializers.Serializers;
 import com.artillexstudios.axapi.utils.StringUtils;
-import com.artillexstudios.axvaults.converters.ItemReplacer;
 import com.artillexstudios.axvaults.database.Database;
 import com.artillexstudios.axvaults.placed.PlacedVaults;
 import com.artillexstudios.axvaults.utils.ThreadUtils;
@@ -26,7 +25,6 @@ import java.sql.Statement;
 import java.util.*;
 
 import static com.artillexstudios.axvaults.AxVaults.CONFIG;
-import static com.artillexstudios.axvaults.converters.ItemReplacer.loadRules;
 
 public class MySQL implements Database {
     private HikariDataSource dataSource;
@@ -241,66 +239,6 @@ public class MySQL implements Database {
             ex.printStackTrace();
         }
     }
-
-
-    @Override
-    public int replaceItemsInVaults() {
-        int updatedVaults = 0;
-
-        final String sql = "SELECT * FROM axvaults_data;";
-        final String updateSql = "UPDATE axvaults_data SET storage = ? WHERE uuid = ? AND id = ?;";
-
-        final List<ItemReplacer.ReplacementRule> replacementRules = loadRules();
-
-        try (Connection conn = dataSource.getConnection(); PreparedStatement stmt = conn.prepareStatement(sql); ResultSet rs = stmt.executeQuery()) {
-            while (rs.next()) {
-
-                final UUID uuid = UUID.fromString(rs.getString(2));
-                final int id = rs.getInt(1);
-
-                final ItemStack[] items;
-                try {
-                    items = Serializers.ITEM_ARRAY.deserialize(rs.getBytes(3));
-                } catch (Exception ex) {
-                    ex.printStackTrace();
-                    continue;
-                }
-
-                final ItemStack[] applyResult = ItemReplacer.apply(items, replacementRules, Bukkit.getOfflinePlayer(uuid).getName());
-                if (Arrays.equals(items, applyResult)) continue;
-
-                final byte[] serialized;
-                try {
-                    serialized = Serializers.ITEM_ARRAY.serialize(applyResult);
-                } catch (Exception ex) {
-                    ex.printStackTrace();
-                    continue;
-                }
-
-                try (PreparedStatement updateStmt = conn.prepareStatement(updateSql)) {
-                    updateStmt.setBytes(1, serialized);
-                    updateStmt.setString(2, uuid.toString());
-                    updateStmt.setInt(3, id);
-                    updateStmt.executeUpdate();
-                }
-
-                updatedVaults++;
-
-                final VaultPlayer vaultPlayer = VaultManager.getPlayers().get(uuid);
-                if (vaultPlayer == null) continue;
-
-                final Vault vault = vaultPlayer.getVault(id);
-                if (vault == null) continue;
-
-                ThreadUtils.runSync(() -> vault.setContents(applyResult));
-            }
-        } catch (SQLException ex) {
-            ex.printStackTrace();
-        }
-
-        return updatedVaults;
-    }
-
 
     @Override
     public Set<UUID> getVaultOwners() {
